# src/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Kubernetes Namespace Provisioner API
  description: |
    TypeScript implementation of the Eclipse Che server's Kubernetes namespace provisioning API.
    
    This API provides endpoints for provisioning and managing Kubernetes namespaces where 
    authenticated users can create workspaces.
    
    ## Original Java Implementation
    This API is based on the Eclipse Che server endpoint:
    - Java Class: `org.eclipse.che.workspace.infrastructure.kubernetes.api.server.KubernetesNamespaceService`
    - Endpoint: `POST /kubernetes/namespace/provision`
    
    ## Authentication
    The API supports two authentication methods:
    1. **Bearer Token**: Format `userid:username` (e.g., `Bearer user123:johndoe`)
    2. **Basic Auth**: Username and userid as password (e.g., `johndoe:user123`)
    
    ## Features
    - ✅ Provision Kubernetes namespaces with configurable name templates
    - ✅ List all managed namespaces
    - ✅ Automatic namespace creation if not exists
    - ✅ Namespace metadata with status and attributes
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Eclipse Public License 2.0
    url: https://www.eclipse.org/legal/epl-2.0/
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /factory/resolver:
    post:
      tags:
        - factory
      summary: Resolve factory from parameters
      description: |
        Create factory by providing map of parameters. Returns JSON with factory information.
        
        This endpoint resolves a factory from various sources:
        
        ### Supported URL Types
        - **Raw devfile URLs**: Direct links to devfile.yaml or .devfile.yaml files
        - **Git repository URLs**: Repository URLs (with or without .git suffix)
        - **Git clone URLs**: URLs ending with .git (e.g., `https://github.com/user/repo.git`)
        - **Branch-specific URLs**: URLs pointing to specific branches
        - **SCM providers**: GitHub, GitLab, Bitbucket, Azure DevOps, and GitHub Enterprise
        
        ### Automatic Devfile Detection
        For Git repository URLs, the resolver automatically searches for devfile in this order:
        1. `devfile.yaml`
        2. `.devfile.yaml`
        
        You can configure additional filenames via `CHE_FACTORY_DEFAULT_DEVFILE_FILENAMES` environment variable.
        
        **Note**: This is a beta feature that may be significantly changed.
        
        ## How it works
        1. Accepts parameters (primarily a URL)
        2. Detects URL type (raw devfile or repository)
        3. For repository URLs, strips .git suffix and searches for devfile
        4. Fetches devfile content from the detected location
        5. Creates factory with devfile v2 structure
        6. Returns factory metadata with links
        
        ## Validation
        Use `validate=true` query parameter to validate the factory according to acceptance rules.
        
        ## Examples
        ```
        # GitHub with .git
        {"url": "https://github.com/eclipse-che/che-dashboard.git"}
        
        # GitHub without .git
        {"url": "https://github.com/eclipse-che/che-server"}
        
        # With branch
        {"url": "https://github.com/user/repo/tree/develop"}
        
        # Raw devfile
        {"url": "https://raw.githubusercontent.com/user/repo/main/devfile.yaml"}
        ```
      operationId: resolveFactory
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: validate
          in: query
          description: Whether or not to validate values like it is done when accepting a Factory
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        description: |
          Parameters provided to create factory (typically containing URL).
          
          The URL can be:
          - A direct devfile URL (raw file URL)
          - A Git repository URL (automatically searches for devfile.yaml or .devfile.yaml)
          - A Git clone URL ending with .git (automatically searches for devfile)
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: |
                    URL pointing to a devfile or Git repository. Supports:
                    - Direct devfile URLs: https://raw.githubusercontent.com/user/repo/main/devfile.yaml
                    - Repository URLs: https://github.com/user/repo
                    - Clone URLs: https://github.com/user/repo.git
                    - Branch-specific: https://github.com/user/repo/tree/branch
                    
                    For repository URLs, automatically tries: devfile.yaml, .devfile.yaml
                  example: https://github.com/eclipse-che/che-dashboard.git
              additionalProperties:
                type: string
            examples:
              gitHubCloneUrl:
                summary: GitHub repository with .git suffix
                description: Automatically searches for devfile.yaml or .devfile.yaml in the repository
                value:
                  url: https://github.com/eclipse-che/che-dashboard.git
              gitHubRepoUrl:
                summary: GitHub repository URL
                description: Repository URL without .git suffix
                value:
                  url: https://github.com/eclipse-che/che-server
              gitHubBranchUrl:
                summary: GitHub repository with branch
                description: Repository URL with specific branch
                value:
                  url: https://github.com/user/repo/tree/develop
              gitLabRepoUrl:
                summary: GitLab repository URL
                description: GitLab repository with .git suffix
                value:
                  url: https://gitlab.com/eclipse-che/che-dashboard.git
              bitbucketRepoUrl:
                summary: Bitbucket repository URL
                description: Bitbucket repository URL
                value:
                  url: https://bitbucket.org/user/repo.git
              rawDevfileUrl:
                summary: Direct devfile URL
                description: Direct link to a devfile.yaml file
                value:
                  url: https://raw.githubusercontent.com/eclipse-che/che-server/main/devfile.yaml
              hiddenDevfileUrl:
                summary: Hidden devfile URL
                description: Direct link to a .devfile.yaml file
                value:
                  url: https://raw.githubusercontent.com/user/repo/main/.devfile.yaml
      responses:
        '200':
          description: Factory successfully built from parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryMeta'
              examples:
                gitHubFactory:
                  summary: Factory from GitHub repository
                  description: Factory created from GitHub repository URL with .git suffix
                  value:
                    v: '4.0'
                    name: che-dashboard
                    source: https://github.com/eclipse-che/che-dashboard.git
                    devfile:
                      schemaVersion: '2.1.0'
                      metadata:
                        name: che-dashboard
                        description: Eclipse Che Dashboard
                      projects:
                        - name: che-dashboard
                          git:
                            remotes:
                              origin: https://github.com/eclipse-che/che-dashboard.git
                    links:
                      - rel: self
                        href: /factory/resolver?url=https://github.com/eclipse-che/che-dashboard.git
                gitLabFactory:
                  summary: Factory from GitLab repository
                  description: Factory created from GitLab repository URL
                  value:
                    v: '4.0'
                    name: che-dashboard
                    source: https://gitlab.com/eclipse-che/che-dashboard.git
                    devfile:
                      schemaVersion: '2.1.0'
                      metadata:
                        name: che-dashboard
                    links:
                      - rel: self
                        href: /factory/resolver?url=https://gitlab.com/eclipse-che/che-dashboard.git
                rawDevfileFactory:
                  summary: Factory from raw devfile URL
                  description: Factory created from direct devfile.yaml URL
                  value:
                    v: '4.0'
                    name: che-server
                    source: devfile.yaml
                    devfile:
                      schemaVersion: '2.1.0'
                      metadata:
                        name: che-server
                    links:
                      - rel: self
                        href: /factory/resolver?url=https://raw.githubusercontent.com/eclipse-che/che-server/main/devfile.yaml
        '400':
          description: Missed required parameters or failed to validate factory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                noParameters:
                  value:
                    error: Bad Request
                    message: Factory build parameters required
                notResolvable:
                  value:
                    error: Bad Request
                    message: Cannot build factory with any of the provided parameters. Please check parameters correctness, and resend query.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /factory/token/refresh:
    post:
      tags:
        - factory
      summary: Refresh factory OAuth token
      description: |
        Validate the factory related OAuth token and update/create it if needed.
        
        This endpoint checks and refreshes personal access tokens for SCM providers
        used to access factory sources (GitHub, GitLab, Bitbucket, etc.).
        
        The token refresh process:
        1. Parses the factory URL to determine SCM provider
        2. Checks if authorization was previously rejected
        3. If not rejected, validates or refreshes the token
        4. Stores updated token securely
        
        **Note**: Set `CHE_FORCE_REFRESH_PERSONAL_ACCESS_TOKEN=true` environment variable
        to force token refresh even if existing token is valid.
      operationId: refreshFactoryToken
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: url
          in: query
          description: Factory URL for which to refresh the OAuth token
          required: true
          schema:
            type: string
          example: https://github.com/user/repo
      responses:
        '200':
          description: The factory related OAuth token is valid or has been updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token refreshed successfully
        '400':
          description: Factory URL is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Bad Request
                message: Factory url required
        '401':
          description: Failed to update the factory related OAuth token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Unauthorized
                message: Failed to update the factory related OAuth token
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    FactoryMeta:
      type: object
      required:
        - v
      properties:
        v:
          type: string
          description: Factory version
          example: '4.0'
        name:
          type: string
          description: Factory name
          example: my-workspace
        id:
          type: string
          description: Factory ID (set by server)
          example: factory123
        source:
          type: string
          description: |
            Indicates filename in repository from which the factory was created 
            (for example, devfile.yaml) or just contains 'repo' value if factory 
            was created from bare repository. For custom raw URLs (pastebin, gist etc) 
            value is null
          example: devfile.yaml
        devfile:
          type: object
          description: |
            Devfile v2 as a generic object. Since che-server doesn't know the structure 
            of Devfile v2, we use a generic object type.
          additionalProperties: true
          example:
            schemaVersion: '2.1.0'
            metadata:
              name: my-project
        scm_info:
          type: object
          description: SCM (Source Control Management) information
          properties:
            cloneUrl:
              type: string
              example: https://github.com/user/repo.git
            branch:
              type: string
              example: main
            scmProviderName:
              type: string
              example: github
            scmProviderUrl:
              type: string
              example: https://github.com
        creator:
          type: object
          description: Factory creator information
          properties:
            userId:
              type: string
              example: user123
            userName:
              type: string
              example: johndoe
            created:
              type: integer
              format: int64
              description: Creation timestamp
              example: 1699876543000
        policies:
          type: object
          description: Factory policies
          properties:
            referer:
              type: string
            since:
              type: integer
              format: int64
            until:
              type: integer
              format: int64
        ide:
          type: object
          description: IDE configuration
          properties:
            onAppClosed:
              type: string
            onAppLoaded:
              type: string
            onProjectsLoaded:
              type: string
        links:
          type: array
          description: HAL links for factory operations
          items:
            type: object
            properties:
              rel:
                type: string
                example: self
              href:
                type: string
                example: /factory/resolver?url=https://example.com/devfile.yaml
              method:
                type: string
                example: GET
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type or category
          example: Unauthorized
        message:
          type: string
          description: Detailed error message
          example: Authorization header is required
        details:
          type: string
          description: Additional error details (only in development mode)
          example: Missing Authorization header in request
    ExtendedError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            stack:
              type: string
              description: Stack trace (only in development mode)
              example: 'Error: Not able to find namespace\n    at NamespaceProvisioner.provision...'
  responses:
    Unauthorized:
      description: Unauthorized - Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingAuth:
              value:
                error: Unauthorized
                message: Authorization header is required
            invalidAuth:
              value:
                error: Unauthorized
                message: Invalid authorization credentials
    InternalServerError:
      description: Internal server error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExtendedError'
          examples:
            namespaceNotFound:
              value:
                error: Internal Server Error
                message: Not able to find namespace che-johndoe
            kubernetesError:
              value:
                error: Internal Server Error
                message: Failed to create namespace
                details: Kubernetes API error - Forbidden
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: userid:username
      description: |
        Bearer token authentication using the format `userid:username`.
        
        Example: `Authorization: Bearer user123:johndoe`
        
        In production, this should be replaced with proper JWT tokens.

    BasicAuth:
      type: http
      scheme: basic
      description: |
        Basic authentication using username and userid.
        
        - Username: user's username (e.g., `johndoe`)
        - Password: user's ID (e.g., `user123`)
        
        Example: `Authorization: Basic am9obmRvZTp1c2VyMTIz` (base64 of `johndoe:user123`)
security:
  - BearerAuth: []
  - BasicAuth: []